requirements.txt
cssselect2==0.7.0
cx_Oracle==8.3.0
decorator==4.4.2
Django==5.0.3
django-formtools==2.5.1
django-otp==1.5.0
django-timezone-field==6.1.0
hyperlink==21.0.0
html5lib==1.1
openpyxl==3.1.2
proglog==0.1.10
pypdf==4.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-docx==1.1.0
reportlab==4.2.0
six==1.16.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.16.0
xhtml2pdf==0.2.16
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.4.post1
psycopg2
